requirements.txt
airflow==0.6
boto3==1.26.104
botocore==1.29.104
dag_factory==0.17.3
nltk==3.8.1
pandas==1.5.3
python-decouple==3.8
SQLAlchemy==1.4.47
